+++ CHANGELOG.md
@@ -2929,3 +2929,4 @@
 - [2026-07-17 22:57] fix(storage): fix memory leak by cancelling storage subscription
 - [2026-07-18 09:00] feat(cache): implement shimmer loading skeleton in cache
 - [2026-07-18 09:33] docs(navigation): update readme and API contracts for navigation
+- [2026-07-18 14:04] feat(i18n): integrate offline caching for i18n state